feat: consolidate link-related field types (#114)
* feat: consolidate link-related field types

Merge `link-to-media` into `link` with a new `--allow` option that
accepts `document`, `media`, or `web`. Omitting `--allow` creates a
generic link. Keep `content-relationship` separate.

Update descriptions on `link` and `content-relationship` so agents
can tell when to use each one.


* fix: validate --allow option and restore field-edit subtype guards

Validate that --allow is one of document, media, or web. Restore
subtype guards in field-edit so content-relationship fields can't
receive link-only options and vice versa.


* fix: allow all Link options regardless of select value

Remove the select-based guard in field-edit so link fields with
select: "document" can still be edited with link-specific options
like --allow-target-blank and --allow-text.


---------
