+++ src/commands/field-add-link.ts
@@ -9,13 +9,18 @@ import { getRepositoryName } from "../project";
 
 const config = {
 	name: "prismic field add link",
-	description: "Add a link field to a slice or custom type.",
+	description:
+		"Add a link field to a slice or custom type. Use for navigational links to URLs, documents, or media. For data-level relations between documents, use content-relationship instead.",
 	positionals: {
 		id: { description: "Field ID", required: true },
 	},
 	options: {
 		...TARGET_OPTIONS,
 		label: { type: "string", description: "Field label" },
+		allow: {
+			type: "string",
+			description: "Restrict to a link type: document, media, or web",
+		},
 		"allow-target-blank": { type: "boolean", description: "Allow opening in new tab" },
 		"allow-text": { type: "boolean", description: "Allow custom link text" },
 		repeatable: { type: "boolean", description: "Allow multiple links" },
@@ -25,14 +30,22 @@ const config = {
 
 export default createCommand(config, async ({ positionals, values }) => {
 	const [id] = positionals;
+	const ALLOWED_LINK_TYPES = ["document", "media", "web"] as const;
+
 	const {
 		label,
+		allow,
 		"allow-target-blank": allowTargetBlank,
 		"allow-text": allowText,
 		repeatable: repeat,
 		variant: variants,
 		repo = await getRepositoryName(),
 	} = values;
+
+	if (allow && !ALLOWED_LINK_TYPES.includes(allow as (typeof ALLOWED_LINK_TYPES)[number])) {
+		throw new CommandError(`--allow must be one of: ${ALLOWED_LINK_TYPES.join(", ")}`);
+	}
+	const select = allow as (typeof ALLOWED_LINK_TYPES)[number] | undefined;
 
 	const token = await getToken();
 	const host = await getHost();
@@ -43,6 +56,7 @@ export default createCommand(config, async ({ positionals, values }) => {
 		type: "Link",
 		config: {
 			label: label ?? capitalCase(fieldId),
+			select,
 			allowTargetBlank,
 			allowText,
 			repeat,
