+++ test/field-add-link.test.ts
@@ -1,5 +1,10 @@
 import { buildCustomType, buildSlice, it } from "./it";
-import { getCustomTypes, getSlices, insertCustomType, insertSlice } from "./prismic";
+import {
+	getCustomTypes,
+	getSlices,
+	insertCustomType,
+	insertSlice,
+} from "./prismic";
 
 it("supports --help", async ({ expect, prismic }) => {
 	const { stdout, exitCode } = await prismic("field", ["add", "link", "--help"]);
@@ -45,4 +50,26 @@ it("adds a link field to a custom type", async ({ expect, prismic, repo, token, 
 	const updated = customTypes.find((ct) => ct.id === customType.id);
 	const field = updated!.json.Main.my_link;
 	expect(field).toMatchObject({ type: "Link" });
+});
+
+it("adds a media link field with --allow media", async ({ expect, prismic, repo, token, host }) => {
+	const slice = buildSlice();
+	await insertSlice(slice, { repo, token, host });
+
+	const { stdout, exitCode } = await prismic("field", [
+		"add",
+		"link",
+		"my_media",
+		"--to-slice",
+		slice.name,
+		"--allow",
+		"media",
+	]);
+	expect(exitCode).toBe(0);
+	expect(stdout).toContain("Field added: my_media");
+
+	const slices = await getSlices({ repo, token, host });
+	const updated = slices.find((s) => s.id === slice.id);
+	const field = updated!.variations[0].primary!.my_media;
+	expect(field).toMatchObject({ type: "Link", config: { select: "media" } });
 });
